+++ telemetry/aggregator/src/main/resources/application.yml
@@ -1,0 +1,29 @@
+#grpc:
+#  server:
+#    port: 59091
+
+
+#server:
+#  port: 8080
+
+logging:
+  level:
+    ru.practicum: trace
+
+spring:
+ kafka:
+   bootstrap-servers: localhost:9092
+   topics:
+    telemetry-hubs-topics: "telemetry.hubs.v1"
+    telemetry-sensors-topics: "telemetry.sensors.v1"
+    telemetry-snapshots-topics: "telemetry.snapshots.v1"
+   consumer:
+    client-id: "SomeConsumer"
+    group-id: "some.group.id-"
+    auto-offset-reset: "earliest"
+    key-deserializer: "org.apache.kafka.common.serialization.StringDeserializer"
+    value-deserializer: "org.apache.kafka.common.serialization.StringDeserializer"
+    max-poll-record-config: 100
+    fetch-max-bytes-config: 3072000
+    max-partition-fetch-bytes-config: 307200
+
