+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/service/handler/hub/DeviceRemovedHubEventHandler.java
@@ -4,14 +4,14 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.grpc.telemetry.event.DeviceRemovedEventProto;
 import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
 import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
-import ru.yandex.practicum.telemetry.collector.service.KafkaClient;
+import ru.yandex.practicum.telemetry.collector.service.KafkaEventProducer;
 import ru.yandex.practicum.telemetry.collector.service.handler.HubEventHandler;
 
 @Component
 public class DeviceRemovedHubEventHandler extends BaseHubEventHandler<DeviceRemovedEventAvro> implements HubEventHandler {
 
-      public DeviceRemovedHubEventHandler(KafkaClient kafkaClient) {
-        super(kafkaClient);
+      public DeviceRemovedHubEventHandler(KafkaEventProducer kafkaEventProducer) {
+        super(kafkaEventProducer);
     }
 
     @Override
