+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/service/handler/sensor/ClimateSensorEventHandler.java
@@ -4,13 +4,13 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.grpc.telemetry.event.ClimateSensorProto;
 import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
 import ru.yandex.practicum.kafka.telemetry.event.ClimateSensorAvro;
-import ru.yandex.practicum.telemetry.collector.service.KafkaClient;
+import ru.yandex.practicum.telemetry.collector.service.KafkaEventProducer;
 import ru.yandex.practicum.telemetry.collector.service.handler.SensorEventHandler;
 
 @Component
 public class ClimateSensorEventHandler extends BaseSensorEventHandler<ClimateSensorAvro> implements SensorEventHandler {
 
-    public ClimateSensorEventHandler(KafkaClient producer) {
+    public ClimateSensorEventHandler(KafkaEventProducer producer) {
         super(producer);
     }
