+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/service/handler/sensor/LightSensorEventHandler.java
@@ -4,13 +4,13 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.grpc.telemetry.event.LightSensorProto;
 import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
 import ru.yandex.practicum.kafka.telemetry.event.LightSensorAvro;
-import ru.yandex.practicum.telemetry.collector.service.KafkaClient;
+import ru.yandex.practicum.telemetry.collector.service.KafkaEventProducer;
 import ru.yandex.practicum.telemetry.collector.service.handler.SensorEventHandler;
 
 @Component
 public class LightSensorEventHandler extends BaseSensorEventHandler<LightSensorAvro> implements SensorEventHandler {
 
-    public LightSensorEventHandler(KafkaClient producer) {
+    public LightSensorEventHandler(KafkaEventProducer producer) {
         super(producer);
     }
