+++ telemetry/collector/src/main/resources/application.yml
@@ -16,14 +16,3 @@ spring:
    topics:
     telemetry-hubs-topics: "telemetry.hubs.v1"
     telemetry-sensors-topics: "telemetry.sensors.v1"
-    telemetry-snapshots-topics: "telemetry.snapshots.v1"
-   consumer:
-    client-id: "SomeConsumer"
-    group-id: "some.group.id-"
-    auto-offset-reset: "earliest"
-    key-deserializer: "org.apache.kafka.common.serialization.StringDeserializer"
-    value-deserializer: "org.apache.kafka.common.serialization.StringDeserializer"
-    max-poll-record-config: 100
-    fetch-max-bytes-config: 3072000
-    max-partition-fetch-bytes-config: 307200
-
